+++ util/src/main/java/com/google/cloud/hadoop/util/testing/MockHttpTransportHelper.java
@@ -1,3 +1,17 @@
+/*
+ * Copyright 2020 Google LLC. All Rights Reserved.
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License. You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software distributed under the
+ * License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+ * express or implied. See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package com.google.cloud.hadoop.util.testing;
 
 import static com.google.api.client.http.HttpStatusCodes.STATUS_CODE_NOT_FOUND;
@@ -27,16 +41,18 @@ import com.google.api.client.testing.http.HttpTesting;
 import com.google.api.client.testing.http.MockHttpTransport;
 import com.google.api.client.testing.http.MockLowLevelHttpRequest;
 import com.google.api.client.testing.http.MockLowLevelHttpResponse;
+import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.io.CharStreams;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
-import java.util.Collections;
 import java.util.Map;
 
+/** Utility methods for testing with {@link MockHttpTransport} */
 public final class MockHttpTransportHelper {
 
+  /** HTTP JSON API error responses */
   public enum ErrorResponses {
     GONE(SC_GONE, STATUS_CODE_SERVICE_UNAVAILABLE, "backendError", "Backend Error", GLOBAL_DOMAIN),
     NOT_FOUND(STATUS_CODE_NOT_FOUND, "notFound", "Not Found", GLOBAL_DOMAIN),
@@ -218,7 +234,7 @@ public final class MockHttpTransportHelper {
 
     GoogleJsonError jsonError = new GoogleJsonError();
     jsonError.setCode(errorResponse.getErrorCode());
-    jsonError.setErrors(Collections.singletonList(errorInfo));
+    jsonError.setErrors(ImmutableList.of(errorInfo));
     jsonError.setMessage(errorResponse.getErrorMessage());
     jsonError.setFactory(JSON_FACTORY);