+++ bigquery/CHANGES.md
@@ -8,7 +8,7 @@
 
 1.  Remove Log4j backend for Google Flogger.
 
-1.  Add properties to override Google Cloud API endpoints: 
+1.  Add properties to override Google Cloud API endpoints:
 
         mapred.bq.bigquery.root.url
         mapred.bq.token.server.url